fix: queue errors

See merge request s10-final/S10P31A401!194

+++ src/main/src/main/java/com/qqueueing/main/registration/model/GetWaitingInfoResDto.java
@@ -5,10 +5,10 @@ import lombok.Setter;
 
 @Getter @Setter
 public class GetWaitingInfoResDto {
-    private int enterCnt;
+    private long enterCnt;
     private int totalQueueSize;
 
-    public GetWaitingInfoResDto(int enterCnt, int totalQueueSize) {
+    public GetWaitingInfoResDto(long enterCnt, int totalQueueSize) {
         this.enterCnt = enterCnt;
         this.totalQueueSize = totalQueueSize;
     }
